Fix error in handling single world replacements.

+++ lib/snippet_generator.rb
@@ -14,8 +14,12 @@ module SnippetGenerator
       def single_snippet_export(file_to_write,trigger,replacement)
         File.open(file_to_write,"a") { |file| file.write(NEW_LINE+'  - trigger: '+'":'+trigger+QUOTE+NEW_LINE) }
         File.open(file_to_write,"a") { |file| file.write('    replace: |-'+NEW_LINE) }
-        replacement.each do |item|
-          File.open(file_to_write,"a") { |file| file.write('           '+item) }
+        if (replacement.instance_of?(String) ) then
+          File.open(file_to_write,"a") { |file| file.write('           '+replacement) }
+        else
+          replacement.each do |item|
+            File.open(file_to_write,"a") { |file| file.write('           '+item) }
+          end
         end
 
       end
